# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koumbaya Marketplace API",
        "description": "API REST pour la marketplace Koumbaya avec système de tombolas et paiements mobile money",
        "contact": {
            "name": "Équipe Koumbaya",
            "email": "dev@koumbaya.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement Koumbaya"
        }
    ],
    "paths": {
...
        "/api/lotteries": {
            "get": {
                "tags": [
                    "Lotteries"
                ],
                "summary": "Lister toutes les tombolas",
                "operationId": "9ea87e50016fca92c67526b0b1c981e0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "active",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrer par catégorie de produit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_ticket_price",
                        "in": "query",
                        "description": "Prix minimum du ticket",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_ticket_price",
                        "in": "query",
                        "description": "Prix maximum du ticket",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "ending_soon",
                        "in": "query",
                        "description": "Tombolas se terminant dans les 24h",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Trier par",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "end_date_asc",
                                "end_date_desc",
                                "ticket_price_asc",
                                "ticket_price_desc",
                                "popularity"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des tombolas"
                    }
                }
            }
        },
...
        "/api/lotteries/{id}": {
            "get": {
                "tags": [
                    "Lotteries"
                ],
                "summary": "Détails d'une tombola",
                "operationId": "fbb03768631c4ce3881e829c0b88da49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tombola",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la tombola"
                    }
                }
            }
        },
        "/api/lotteries/{id}/buy-ticket": {
            "post": {
                "tags": [
                    "Lotteries"
                ],
                "summary": "Acheter un ticket de tombola",
                "operationId": "ce2ff44ef1aca23589f45b6cd481b1a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tombola",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "description": "Nombre de tickets",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket(s) acheté(s)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lotteries/{id}/my-tickets": {
            "get": {
                "tags": [
                    "Lotteries"
                ],
                "summary": "Mes tickets pour une tombola",
                "operationId": "91ac3d540c9f89d679fe367eeb01b08e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tombola",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mes tickets"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lotteries/{id}/draw": {
            "post": {
                "tags": [
                    "Lotteries"
                ],
                "summary": "Effectuer le tirage d'une tombola (Admin/Merchant seulement)",
                "operationId": "a1e7277d52ec67123ce470184b911246",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tombola",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tirage effectué"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Bearer JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}